Move db module from sats to lib
Update bindings crate

Update table crate

Update vm crate

Update bindings-macro crate

Update rust-wasm-test module

Update core crate

Update cli crate

Update bench crate

Remove needless dependency

Placate clippy

Cargo fmt

Placate clippy once more

Back, foul beast

Fix compile error
