+++ crates/core/src/db/datastore/locking_tx_datastore/tx.rs
@@ -6,8 +6,9 @@ use super::{
     SharedReadGuard,
 };
 use crate::execution_context::ExecutionContext;
+use spacetimedb_lib::db::def::TableSchema;
 use spacetimedb_primitives::{ColList, TableId};
-use spacetimedb_sats::{db::def::TableSchema, AlgebraicValue};
+use spacetimedb_sats::AlgebraicValue;
 use std::sync::Arc;
 use std::{
     ops::RangeBounds,
