+++ crates/core/src/error.rs
@@ -14,13 +14,13 @@ use crate::client::ClientActorId;
 use crate::db::datastore::system_tables::SystemTable;
 use crate::host::scheduler::ScheduleError;
 use spacetimedb_lib::buffer::DecodeError;
+use spacetimedb_lib::db::def::IndexDef;
+use spacetimedb_lib::db::error::{LibError, RelationError, SchemaErrors};
+use spacetimedb_lib::relation::FieldName;
 use spacetimedb_lib::ProductValue;
 use spacetimedb_primitives::*;
-use spacetimedb_sats::db::def::IndexDef;
-use spacetimedb_sats::db::error::{LibError, RelationError, SchemaErrors};
 use spacetimedb_sats::hash::Hash;
 use spacetimedb_sats::product_value::InvalidFieldError;
-use spacetimedb_sats::relation::FieldName;
 use spacetimedb_sats::satn::Satn;
 use spacetimedb_vm::errors::{ErrorKind, ErrorLang, ErrorType, ErrorVm};
 use spacetimedb_vm::expr::Crud;
