+++ crates/core/src/host/scheduler.rs
@@ -7,10 +7,10 @@ use rustc_hash::FxHashMap;
 use sled::transaction::TransactionError;
 use spacetimedb_client_api_messages::timestamp::Timestamp;
 use spacetimedb_lib::bsatn::ser::BsatnError;
+use spacetimedb_lib::db::def::TableSchema;
 use spacetimedb_lib::scheduler::ScheduleAt;
 use spacetimedb_lib::Address;
 use spacetimedb_primitives::TableId;
-use spacetimedb_sats::db::def::TableSchema;
 use spacetimedb_sats::{AlgebraicValue, ProductValue};
 use spacetimedb_table::table::RowRef;
 use tokio::sync::mpsc;
