+++ crates/core/src/sql/ast.rs
@@ -1,10 +1,10 @@
 use crate::db::relational_db::{MutTx, RelationalDB, Tx};
 use crate::error::{DBError, PlanError};
 use spacetimedb_data_structures::map::{HashCollectionExt as _, IntMap};
+use spacetimedb_lib::db::def::{ColumnSchema, TableSchema};
+use spacetimedb_lib::db::error::RelationError;
+use spacetimedb_lib::relation::{ColExpr, FieldName};
 use spacetimedb_primitives::ColId;
-use spacetimedb_sats::db::def::{ColumnSchema, TableSchema};
-use spacetimedb_sats::db::error::RelationError;
-use spacetimedb_sats::relation::{ColExpr, FieldName};
 use spacetimedb_sats::{AlgebraicType, AlgebraicValue};
 use spacetimedb_vm::errors::ErrorVm;
 use spacetimedb_vm::expr::{Expr, FieldExpr, FieldOp};
