+++ crates/core/src/subscription/query.rs
@@ -101,13 +101,13 @@ mod tests {
     use crate::vm::DbProgram;
     use itertools::Itertools;
     use spacetimedb_lib::bsatn::to_vec;
+    use spacetimedb_lib::db::auth::{StAccess, StTableType};
+    use spacetimedb_lib::db::def::*;
     use spacetimedb_lib::error::ResultTest;
     use spacetimedb_lib::identity::AuthCtx;
+    use spacetimedb_lib::relation::FieldName;
     use spacetimedb_lib::Identity;
     use spacetimedb_primitives::{ColId, TableId};
-    use spacetimedb_sats::db::auth::{StAccess, StTableType};
-    use spacetimedb_sats::db::def::*;
-    use spacetimedb_sats::relation::FieldName;
     use spacetimedb_sats::{product, AlgebraicType, ProductType, ProductValue};
     use spacetimedb_vm::eval::run_ast;
     use spacetimedb_vm::eval::test_helpers::{mem_table, mem_table_without_table_name, scalar};
