+++ crates/core/src/vm.rs
@@ -10,10 +10,10 @@ use crate::execution_context::ExecutionContext;
 use core::ops::{Bound, RangeBounds};
 use itertools::Itertools;
 use spacetimedb_data_structures::map::IntMap;
+use spacetimedb_lib::db::def::TableDef;
 use spacetimedb_lib::identity::AuthCtx;
+use spacetimedb_lib::relation::{ColExpr, DbTable};
 use spacetimedb_primitives::*;
-use spacetimedb_sats::db::def::TableDef;
-use spacetimedb_sats::relation::{ColExpr, DbTable};
 use spacetimedb_sats::{AlgebraicValue, ProductValue};
 use spacetimedb_table::static_assert_size;
 use spacetimedb_table::table::RowRef;
@@ -623,10 +623,10 @@ pub(crate) mod tests {
     };
     use crate::db::relational_db::tests_utils::TestDB;
     use crate::execution_context::ExecutionContext;
+    use spacetimedb_lib::db::auth::{StAccess, StTableType};
+    use spacetimedb_lib::db::def::{ColumnDef, IndexDef, IndexType, TableSchema};
     use spacetimedb_lib::error::ResultTest;
-    use spacetimedb_sats::db::auth::{StAccess, StTableType};
-    use spacetimedb_sats::db::def::{ColumnDef, IndexDef, IndexType, TableSchema};
-    use spacetimedb_sats::relation::{FieldName, Header};
+    use spacetimedb_lib::relation::{FieldName, Header};
     use spacetimedb_sats::{product, AlgebraicType, ProductType, ProductValue};
     use spacetimedb_vm::eval::run_ast;
     use spacetimedb_vm::eval::test_helpers::{mem_table, mem_table_one_u64, scalar};
