+++ crates/lib/src/db/attr.rs
@@ -1,0 +1,1 @@
+pub use spacetimedb_primitives::ColumnAttribute;
