+++ crates/lib/src/lib.rs
@@ -1,24 +1,31 @@
+use crate::db::def::TableDef;
+use anyhow::Context;
+use sats::typespace::TypespaceBuilder;
+use spacetimedb_sats::{impl_serialize, WithTypespace};
 use std::any::TypeId;
 use std::collections::{btree_map, BTreeMap};
 
-use anyhow::Context;
-use sats::typespace::TypespaceBuilder;
-use spacetimedb_sats::db::def::TableDef;
-use spacetimedb_sats::{impl_serialize, WithTypespace};
-
 pub mod address;
+pub mod db;
 pub mod error;
 pub mod filter;
 pub mod identity;
 pub mod operator;
+pub mod relation;
 pub mod scheduler;
 pub mod version;
+
+pub mod type_def {
+    pub use spacetimedb_sats::{AlgebraicType, ProductType, ProductTypeElement, SumType};
+}
+pub mod type_value {
+    pub use spacetimedb_sats::{AlgebraicValue, ProductValue};
+}
 
 pub use address::Address;
 pub use identity::Identity;
 pub use scheduler::ScheduleAt;
 pub use spacetimedb_sats::hash::{self, hash_bytes, Hash};
-pub use spacetimedb_sats::relation;
 pub use spacetimedb_sats::SpacetimeType;
 pub use spacetimedb_sats::__make_register_reftype;
 pub use spacetimedb_sats::{self as sats, bsatn, buffer, de, ser};
