+++ crates/sats/src/lib.rs
@@ -8,17 +8,16 @@ pub mod bsatn;
 pub mod buffer;
 pub mod builtin_type;
 pub mod convert;
-pub mod db;
 pub mod de;
 pub mod hash;
 pub mod hex;
 pub mod map_type;
 pub mod map_value;
 pub mod meta_type;
+pub mod primitives;
 pub mod product_type;
 pub mod product_type_element;
 pub mod product_value;
-pub mod relation;
 mod resolve_refs;
 pub mod satn;
 pub mod ser;
