+++ crates/table/src/read_column.rs
@@ -340,11 +340,8 @@ mod test {
     use super::*;
     use crate::{blob_store::HashMapBlobStore, indexes::SquashedOffset, table::Table};
     use proptest::{prelude::*, prop_assert_eq, proptest, test_runner::TestCaseResult};
-    use spacetimedb_sats::{
-        db::def::{TableDef, TableSchema},
-        product,
-        proptest::generate_typed_row,
-    };
+    use spacetimedb_lib::db::def::{TableDef, TableSchema};
+    use spacetimedb_sats::{product, proptest::generate_typed_row};
 
     fn table(ty: ProductType) -> Table {
         let def = TableDef::from_product("", ty);
